requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b2
azure-common==1.1.28
azure-core==1.30.1
azure-functions==1.19.0
azure-storage-blob==12.19.1
azure-storage-file-share==12.15.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
frozenlist==1.4.1
idna==3.7
isodate==0.6.1
msrest==0.7.1
multidict==6.0.5
oauthlib==3.2.2
pycparser==2.22
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
